simplify the motion by combing code
lots of repeated code.  Replaced this with a dictionary storing the parameters for each motion type (which were all identical in the code); seemed a bit more modular to make this settings a dictionary rather than provide individual parameters.  Left the up/down/left/right functions in place to avoid breaking things elsewhere